Задание 5. Добавление классы TestBase и перенос методов
Сделала общий базовый класс для тестов TestBase, перенесла вспомогательные методы в него. Добавила тест на удаление контакта.
