+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/ContactDeletionTests.java
@@ -1,0 +1,15 @@
+package ru.stqa.pft.addressbook;
+
+import org.testng.annotations.Test;
+
+public class ContactDeletionTests extends TestBase {
+
+  @Test
+  public void testContactDeletion() throws Exception {
+    selectContact();
+    deleteSelectedContacts();
+    acceptDeletion();
+    returnToHomepage();
+  }
+
+}
