+++ addressbook-web-test/src/test/java/ru/stqa/pft/addressbook/GroupDeletionTests.java
@@ -1,0 +1,15 @@
+package ru.stqa.pft.addressbook;
+
+import org.testng.annotations.Test;
+
+public class GroupDeletionTests extends TestBase{
+
+  @Test
+  public void testGroupDeletion() throws Exception {
+    gotoGroupPage();
+    selectGroup();
+    deleteSelectedGroups();
+    returnToGroupPage();
+  }
+
+}
